backend/requirements.txt
fastapi
uvicorn
websockets
firebase-admin
pydantic
bcrypt
pycryptodome
python-dotenv
gunicorn

 
